+++ packages/soda-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react
 
+## 0.0.23
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.40.5
+
 ## 0.0.22
 
 ### Patch Changes